+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHP41GJS9CCC7Y7D22KX2HCE"
+        "01KHP71YYN4D8T30HG6P779V7M"
     )
     r"""The unique identifier of the data source"""